Reduced db calls for update and softDelete methods.

+++ Technikon-App-BackEnd/src/main/java/com/scytalys/technikon/domain/User.java
@@ -23,6 +23,6 @@ public class User {
     private String password;
     private boolean isActive = true;
     @Version
-    private int version;
+    private long version;
 
 }
